split rrd functions exporters
