+++ database/rrdfunctions-inflight.h
@@ -5,4 +5,12 @@
 
 #include "rrd.h"
 
+void rrd_functions_inflight_init(void);
+
+// cancel a running function, to be run from anywhere
+void rrd_function_cancel(const char *transaction);
+
+void rrd_function_progress(const char *transaction);
+void rrd_function_call_progresser(uuid_t *transaction);
+
 #endif //NETDATA_RRDFUNCTIONS_INFLIGHT_H
