+++ database/rrdfunctions.c
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
+
 #define NETDATA_RRD_INTERNALS
 #define RRD_COLLECTOR_INTERNALS
-
 #define RRD_FUNCTIONS_INTERNALS
 #include "rrd.h"
 
@@ -266,46 +266,6 @@ void rrd_function_add(RRDHOST *host, RRDSET *st, const char *name, int timeout, 
     dictionary_acquired_item_release(host->functions, item);
 }
 
-void rrd_functions_expose_rrdpush(RRDSET *st, BUFFER *wb) {
-    if(!st->functions_view)
-        return;
-
-    struct rrd_host_function *tmp;
-    dfe_start_read(st->functions_view, tmp) {
-        buffer_sprintf(wb
-                       , PLUGINSD_KEYWORD_FUNCTION " \"%s\" %d \"%s\" \"%s\" \"%s\" %d\n"
-                       , tmp_dfe.name
-                       , tmp->timeout
-                       , string2str(tmp->help)
-                       , string2str(tmp->tags)
-                       , http_id2access(tmp->access)
-                       , tmp->priority
-                       );
-    }
-    dfe_done(tmp);
-}
-
-void rrd_functions_expose_global_rrdpush(RRDHOST *host, BUFFER *wb) {
-    rrdhost_flag_clear(host, RRDHOST_FLAG_GLOBAL_FUNCTIONS_UPDATED);
-
-    struct rrd_host_function *tmp;
-    dfe_start_read(host->functions, tmp) {
-        if(!(tmp->options & RRD_FUNCTION_GLOBAL))
-            continue;
-
-        buffer_sprintf(wb
-                , PLUGINSD_KEYWORD_FUNCTION " GLOBAL \"%s\" %d \"%s\" \"%s\" \"%s\" %d\n"
-                , tmp_dfe.name
-                , tmp->timeout
-                , string2str(tmp->help)
-                , string2str(tmp->tags)
-                , http_id2access(tmp->access)
-               , tmp->priority
-        );
-    }
-    dfe_done(tmp);
-}
-
 struct {
     const char *format;
     HTTP_CONTENT_TYPE content_type;
@@ -406,109 +366,3 @@ int rrd_functions_find_by_name(RRDHOST *host, BUFFER *wb, const char *name, size
     return HTTP_RESP_OK;
 }
 
-// ----------------------------------------------------------------------------
-
-static void functions2json(DICTIONARY *functions, BUFFER *wb)
-{
-    struct rrd_host_function *t;
-    dfe_start_read(functions, t)
-    {
-        if (!rrd_collector_running(t->collector))
-            continue;
-
-        buffer_json_member_add_object(wb, t_dfe.name);
-        {
-            buffer_json_member_add_string_or_empty(wb, "help", string2str(t->help));
-            buffer_json_member_add_int64(wb, "timeout", (int64_t) t->timeout);
-
-            char options[65];
-            snprintfz(
-                    options, 64
-                    , "%s%s"
-                    , (t->options & RRD_FUNCTION_LOCAL) ? "LOCAL " : ""
-                    , (t->options & RRD_FUNCTION_GLOBAL) ? "GLOBAL" : ""
-                    );
-
-            buffer_json_member_add_string_or_empty(wb, "options", options);
-            buffer_json_member_add_string_or_empty(wb, "tags", string2str(t->tags));
-            buffer_json_member_add_string(wb, "access", http_id2access(t->access));
-            buffer_json_member_add_uint64(wb, "priority", t->priority);
-        }
-        buffer_json_object_close(wb);
-    }
-    dfe_done(t);
-}
-
-void chart_functions2json(RRDSET *st, BUFFER *wb) {
-    if(!st || !st->functions_view) return;
-
-    functions2json(st->functions_view, wb);
-}
-
-void host_functions2json(RRDHOST *host, BUFFER *wb) {
-    if(!host || !host->functions) return;
-
-    buffer_json_member_add_object(wb, "functions");
-
-    struct rrd_host_function *t;
-    dfe_start_read(host->functions, t) {
-        if(!rrd_collector_running(t->collector)) continue;
-
-        buffer_json_member_add_object(wb, t_dfe.name);
-        {
-            buffer_json_member_add_string(wb, "help", string2str(t->help));
-            buffer_json_member_add_int64(wb, "timeout", t->timeout);
-            buffer_json_member_add_array(wb, "options");
-            {
-                if (t->options & RRD_FUNCTION_GLOBAL)
-                    buffer_json_add_array_item_string(wb, "GLOBAL");
-                if (t->options & RRD_FUNCTION_LOCAL)
-                    buffer_json_add_array_item_string(wb, "LOCAL");
-            }
-            buffer_json_array_close(wb);
-            buffer_json_member_add_string(wb, "tags", string2str(t->tags));
-            buffer_json_member_add_string(wb, "access", http_id2access(t->access));
-            buffer_json_member_add_uint64(wb, "priority", t->priority);
-        }
-        buffer_json_object_close(wb);
-    }
-    dfe_done(t);
-
-    buffer_json_object_close(wb);
-}
-
-void chart_functions_to_dict(DICTIONARY *rrdset_functions_view, DICTIONARY *dst, void *value, size_t value_size) {
-    if(!rrdset_functions_view || !dst) return;
-
-    struct rrd_host_function *t;
-    dfe_start_read(rrdset_functions_view, t) {
-        if(!rrd_collector_running(t->collector)) continue;
-
-        dictionary_set(dst, t_dfe.name, value, value_size);
-    }
-    dfe_done(t);
-}
-
-void host_functions_to_dict(RRDHOST *host, DICTIONARY *dst, void *value, size_t value_size, STRING **help, STRING **tags, HTTP_ACCESS *access, int *priority) {
-    if(!host || !host->functions || !dictionary_entries(host->functions) || !dst) return;
-
-    struct rrd_host_function *t;
-    dfe_start_read(host->functions, t) {
-        if(!rrd_collector_running(t->collector)) continue;
-
-        if(help)
-            *help = t->help;
-
-        if(tags)
-            *tags = t->tags;
-
-        if(access)
-            *access = t->access;
-
-        if(priority)
-            *priority = t->priority;
-
-        dictionary_set(dst, t_dfe.name, value, value_size);
-    }
-    dfe_done(t);
-}
